* Fixed: Compilation problems with MSVC 7.0. Patch from Mike. Doubt: vsnprintf, may be the MSVC version should be lower or for all.

+++ tvision/classes/tinputli.cc
@@ -146,7 +146,7 @@ int TInputLine::mousePos( TEvent& event )
     mouse.x = max( mouse.x, 1 );
     int pos = mouse.x + firstPos - 1;
     pos = max( pos, 0 );
-    pos = min( pos, strlen(data) );
+    pos = min( (size_t)pos, strlen(data) );
     return pos;
 }
